db: separate sql and types
